refactor: move internal proxy functions to functions.go
